+++ src/ball.rs
@@ -2,9 +2,9 @@ use nalgebra::Vector2;
 
 #[derive(Clone, Copy, Debug, PartialEq)]
 pub struct Ball {
-    pub position: Vector2<f32>,
-    pub velocity: Vector2<f32>,
-    pub radius: f32,
-    pub initial_time: f32,
+    pub position: Vector2<f64>,
+    pub velocity: Vector2<f64>,
+    pub radius: f64,
+    pub initial_time: f64,
     pub collision_generation: i32,
 }
